+++ examples/get-started/src/services/telegram/bots/WackzyBot/composers/search.ts
@@ -1,4 +1,0 @@
-import bot from '@botmate/plugin.telegram';
-// const composer = new bot.Composer();
-
-// export default composer;