+++ examples/get-started/src/services/telegram/composers/common.ts
@@ -1,9 +1,0 @@
-/**
- *
- * @param {import('@botmate/plugin.telegram')} telegram
- */
-const composer = ({ telegram }) => {
-  console.log(telegram);
-};
-
-module.exports = composer;