+++ packages/core/helper-plugin/tsup.config.ts
@@ -3,6 +3,7 @@ import { defineConfig } from 'tsup';
 export default defineConfig({
   entry: ['src/index.ts'],
   format: ['cjs', 'esm'],
+  outDir: 'lib',
   splitting: false,
   sourcemap: true,
   clean: true,