+++ examples/get-started/CHANGELOG.md
@@ -4,6 +4,64 @@
 
 ### Patch Changes
 
+- Updated dependencies
+  - @botmate/core@0.0.13
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
+  - @botmate/core@0.0.12
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
+  - @botmate/core@0.0.11
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
+  - @botmate/core@0.0.10
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
+  - @botmate/plugin.telegram@0.0.4
+  - @botmate/core@0.0.9
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
+  - @botmate/core@0.0.8
+  - @botmate/plugin.telegram@0.0.3
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
+  - @botmate/core@0.0.7
+  - @botmate/plugin.telegram@0.0.2
+
+## null
+
+### Patch Changes
+
+- @botmate/core@0.0.6
+
+## null
+
+### Patch Changes
+
 - @botmate/core@0.0.5
 
 ## null