+++ packages/core/admin/CHANGELOG.md
@@ -1,5 +1,46 @@
 # @botmate/admin
 
+## 0.0.7
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @botmate/helper-plugin@0.0.5
+  - @botmate/typescript-utils@0.0.6
+  - @botmate/ui@0.0.5
+  - @botmate/utils@0.0.2
+
+## 0.0.6
+
+### Patch Changes
+
+- fix files and server path
+- Updated dependencies
+  - @botmate/helper-plugin@0.0.4
+  - @botmate/typescript-utils@0.0.5
+  - @botmate/ui@0.0.4
+  - @botmate/utils@0.0.2
+
+## 0.0.5
+
+### Patch Changes
+
+- fix files
+- Updated dependencies
+  - @botmate/helper-plugin@0.0.3
+  - @botmate/typescript-utils@0.0.4
+  - @botmate/ui@0.0.3
+  - @botmate/utils@0.0.2
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @botmate/typescript-utils@0.0.3
+  - @botmate/utils@0.0.2
+
 ## 0.0.3
 
 ### Patch Changes