+++ packages/core/botmate/CHANGELOG.md
@@ -1,5 +1,82 @@
 # @botmate/core
 
+## 0.0.13
+
+### Patch Changes
+
+- fix admin server import
+
+## 0.0.12
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @botmate/typescript-utils@0.0.6
+  - @botmate/plugin.telegram@0.0.5
+  - @botmate/database@0.0.7
+  - @botmate/admin@0.0.7
+  - @botmate/bots@0.0.2
+  - @botmate/utils@0.0.2
+
+## 0.0.11
+
+### Patch Changes
+
+- fix
+
+## 0.0.10
+
+### Patch Changes
+
+- build
+
+## 0.0.9
+
+### Patch Changes
+
+- fix files and server path
+- Updated dependencies
+  - @botmate/typescript-utils@0.0.5
+  - @botmate/plugin.telegram@0.0.4
+  - @botmate/database@0.0.6
+  - @botmate/admin@0.0.6
+  - @botmate/bots@0.0.2
+  - @botmate/utils@0.0.2
+
+## 0.0.8
+
+### Patch Changes
+
+- fix files
+- Updated dependencies
+  - @botmate/typescript-utils@0.0.4
+  - @botmate/database@0.0.5
+  - @botmate/admin@0.0.5
+  - @botmate/bots@0.0.2
+  - @botmate/plugin.telegram@0.0.3
+  - @botmate/utils@0.0.2
+
+## 0.0.7
+
+### Patch Changes
+
+- files fix
+- Updated dependencies
+  - @botmate/typescript-utils@0.0.3
+  - @botmate/database@0.0.4
+  - @botmate/admin@0.0.4
+  - @botmate/bots@0.0.2
+  - @botmate/plugin.telegram@0.0.2
+  - @botmate/utils@0.0.2
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @botmate/database@0.0.3
+
 ## 0.0.5
 
 ### Patch Changes