+++ packages/core/database/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @botmate/database
 
+## 0.0.7
+
+### Patch Changes
+
+- fix
+
+## 0.0.6
+
+### Patch Changes
+
+- fix files and server path
+
+## 0.0.5
+
+### Patch Changes
+
+- fix files
+
+## 0.0.4
+
+### Patch Changes
+
+- files fix
+
+## 0.0.3
+
+### Patch Changes
+
+- fix files
+
 ## 0.0.2
 
 ### Patch Changes