+++ packages/core/helper-plugin/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @botmate/helper-plugin
 
+## 0.0.5
+
+### Patch Changes
+
+- fix
+
+## 0.0.4
+
+### Patch Changes
+
+- fix files and server path
+
+## 0.0.3
+
+### Patch Changes
+
+- fix files
+
 ## 0.0.2
 
 ### Patch Changes