+++ packages/plugins/discord/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @botmate/plugin.discord
 
+## 0.0.3
+
+### Patch Changes
+
+- fix
+
 ## 0.0.2
 
 ### Patch Changes