+++ packages/plugins/telegram/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @botmate/plugin.telegram
 
+## 0.0.5
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @botmate/ui@0.0.5
+  - @botmate/utils@0.0.2
+
+## 0.0.4
+
+### Patch Changes
+
+- fix files and server path
+- Updated dependencies
+  - @botmate/ui@0.0.4
+  - @botmate/utils@0.0.2
+
+## 0.0.3
+
+### Patch Changes
+
+- @botmate/ui@0.0.3
+- @botmate/utils@0.0.2
+
 ## 0.0.2
 
 ### Patch Changes