+++ packages/shared/ui/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @botmate/ui
 
+## 0.0.5
+
+### Patch Changes
+
+- fix
+- Updated dependencies
+  - @botmate/helper-plugin@0.0.5
+
+## 0.0.4
+
+### Patch Changes
+
+- fix files and server path
+- Updated dependencies
+  - @botmate/helper-plugin@0.0.4
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @botmate/helper-plugin@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes